requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-debug-toolbar==6.1.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
docx2pdf==0.1.8
drf-nested-routers==0.95.0
drf-yasg==1.21.11
fire==0.7.1
fonttools==4.61.1
idna==3.11
inflection==0.5.1
lxml==6.0.2
numpy==2.2.6
oauthlib==3.3.1
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2docx==0.5.8
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
PyMuPDF==1.26.7
python-decouple==3.8
python-docx==1.2.0
python3-openid==3.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
termcolor==3.2.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
